Migrate pre-render/eager execution flags to enum
